+++ src/diagnostic.rs
@@ -1,0 +1,5 @@
+use bevy::diagnostic::DiagnosticId;
+
+/// Used to measure render time in miliseconds.
+pub const RENDER_TIME: DiagnosticId =
+    DiagnosticId::from_u128(1187582084072339577959028643519383692);