+++ src/prelude.rs
@@ -1,0 +1,1 @@
+pub use crate::{PixelsOptions, PixelsPlugin, PixelsSet, PixelsWrapper};